requirements.txt
anthropic>=0.8.0
openai>=1.5.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
requests>=2.30.0
python-dotenv>=1.0.0
tqdm>=4.65.0
openpyxl>=3.1.0
concurrent-log-handler>=0.9.20